+++ src/parser/index.ts
@@ -8,13 +8,13 @@ import { parseThisLiteral } from './this-literal';
 import { parseRecordDeclaration } from './record/declaration';
 import { parseRecordReferenceExpr } from './record/reference';
 import { parseVarDeclaration } from './variable-declaration';
-import { parseFunctionDeclaration, parseMethodDeclaration } from './function';
+import { parseFunctionDeclaration } from './function';
 import { parseConditionalExpr } from './conditional';
 import { parseAssignmentExpr } from './assignment';
 import { parseLogicalNotExpr, parseLogicalAndOrExpr } from './logical';
 import { parsePrioritizedExpr } from './prioritized';
 import { parseBinaryOpExpr } from './operation';
-import { parseMethodInvokeExpr } from './method/invocation';
+import { parseMethodDeclaration, parseMethodInvokeExpr } from './method';
 import { parsePipeOperation } from './pipe-operation';
 import { parseLibDirectMethodMapping } from './lib';
 import { ParserError } from './error';
