+++ src/parser/method/index.ts
@@ -1,0 +1,2 @@
+export { parseMethodDeclaration } from './declaration';
+export { parseMethodInvokeExpr } from './invocation';
